requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
openai
PyPDF2
python-docx
langchain[openai]